+++ core-default-node/templates/_Gruntfile.js
@@ -249,7 +249,7 @@ module.exports = function(grunt) {
 			filePathConcatCss:  staticPath+paths.concatCss,
 			filePathsJsNoPrefix:        '',		//will be filled/created in buildfiles task
 			filePathsCssNoPrefix:        '',		//will be filled/created in buildfiles task
-			filePathsJsTestNoPrefix: '',		//for karma .spec files to test
+			filePathsJsTestUnitNoPrefix: '',		//for karma .spec files to test
 			filePathsJsCustom: '',			//for karma coverage files to check
 			filePathsTestProtractor: '',
 			filePathMinJs:      staticPath+paths.minJs,
@@ -284,7 +284,7 @@ module.exports = function(grunt) {
 						outputFiles: {
 							js: ['filePathsJsNoPrefix'],
 							css: ['filePathsCssNoPrefix'],
-							test: ['filePathsJsTestNoPrefix']
+							testUnit: ['filePathsJsTestUnitNoPrefix']
 						}
 					},
 					//index.html file paths (have the static path prefix for use in <link rel="stylesheet" > and <script> tags)
@@ -375,7 +375,7 @@ module.exports = function(grunt) {
 					testProtractor: {
 						moduleGroup: 'testsProtractor',
 						outputFiles: {
-							test: ['filePathsTestProtractor']
+							testE2E: ['filePathsTestProtractor']
 						}
 					}
 				},